Add files via upload
##How to obtain metadata using CoinMarketCap API and more analysis on the pltform

The CMC metadata endpoint returns all the static data that is available for the specified cryptocurrency. The information obtained contains the following: logo, description, website URL, various social links, and others.

In order to obtain crypto metadata, we’ll need to import the CMC library and set the client up.